+++ BaseletElements/src/com/umlet/element/experimental/settings/SettingsUseCase.java
@@ -10,10 +10,10 @@ import com.umlet.element.experimental.settings.facets.DefaultGlobalFacet.Element
 public class SettingsUseCase extends Settings {
 
 	@Override
-	public XValues getXValues(float y, int height, int width) {
-		float b = height/2.0f;
-		float a = width/2.0f;
-		float x = (float) Math.sqrt((1-(Math.pow(b-y, 2) / Math.pow(b, 2)))*Math.pow(a, 2));
+	public XValues getXValues(double y, int height, int width) {
+		double b = height/2.0f;
+		double a = width/2.0f;
+		double x = (double) Math.sqrt((1-(Math.pow(b-y, 2) / Math.pow(b, 2)))*Math.pow(a, 2));
 		return new XValues(a-x, a+x);
 	}
